Move GC scanning logic from FrameIterator to GcScanner
Extract all GC root scanning methods into a dedicated GcScanner class:
- EnumGcRefsForManagedFrame (managed frame via GCInfo)
- GcScanRoots (frame-based dispatch)
- FindGCRefMap (R2R import section resolution)
- PromoteCallerStack/PromoteCallerStackUsingGCRefMap (signature/GCRefMap scanning)
- ScanDynamicHelperFrame (flag-based scanning)

FrameIterator is now purely frame iteration (314 lines, down from 747).
