+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Contracts/Contracts/StackWalk/FrameHandling/FrameIterator.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
-using System.Reflection.Metadata;
-using System.Reflection.Metadata.Ecma335;
 using Microsoft.Diagnostics.DataContractReader.Data;
 
 namespace Microsoft.Diagnostics.DataContractReader.Contracts.StackWalkHelpers;
@@ -312,436 +310,5 @@ internal sealed class FrameIterator
     private static bool InlinedCallFrameHasActiveCall(Data.InlinedCallFrame frame)
     {
         return frame.CallerReturnAddress != TargetPointer.Null;
-    }
-
-    // ===== Frame GC Root Scanning =====
-
-    /// <summary>
-    /// Scans GC roots for a Frame based on its type.
-    /// Dispatches to the appropriate scanning method (GCRefMap, MetaSig, or custom).
-    /// Matches native Frame::GcScanRoots_Impl virtual dispatch.
-    /// </summary>
-    internal void GcScanRoots(TargetPointer frameAddress, GcScanContext scanContext)
-    {
-        if (frameAddress == TargetPointer.Null)
-            return;
-
-        Data.Frame frameData = target.ProcessedData.GetOrAdd<Data.Frame>(frameAddress);
-        FrameType frameType = GetFrameType(target, frameData.Identifier);
-
-        switch (frameType)
-        {
-            case FrameType.StubDispatchFrame:
-            {
-                Data.FramedMethodFrame fmf = target.ProcessedData.GetOrAdd<Data.FramedMethodFrame>(frameAddress);
-                Data.StubDispatchFrame sdf = target.ProcessedData.GetOrAdd<Data.StubDispatchFrame>(frameAddress);
-
-                // Always resolve GCRefMap via FindGCRefMap from the indirection address.
-                TargetPointer gcRefMap = sdf.Indirection != TargetPointer.Null
-                    ? FindGCRefMap(sdf.Indirection)
-                    : TargetPointer.Null;
-
-                if (gcRefMap != TargetPointer.Null)
-                    PromoteCallerStackUsingGCRefMap(fmf.TransitionBlockPtr, gcRefMap, scanContext);
-                else
-                    PromoteCallerStack(frameAddress, fmf.TransitionBlockPtr, scanContext);
-                break;
-            }
-
-            case FrameType.ExternalMethodFrame:
-            {
-                Data.FramedMethodFrame fmf = target.ProcessedData.GetOrAdd<Data.FramedMethodFrame>(frameAddress);
-                Data.ExternalMethodFrame emf = target.ProcessedData.GetOrAdd<Data.ExternalMethodFrame>(frameAddress);
-
-                // Always resolve GCRefMap via FindGCRefMap from the indirection address.
-                TargetPointer gcRefMap = emf.Indirection != TargetPointer.Null
-                    ? FindGCRefMap(emf.Indirection)
-                    : TargetPointer.Null;
-
-                if (gcRefMap != TargetPointer.Null)
-                    PromoteCallerStackUsingGCRefMap(fmf.TransitionBlockPtr, gcRefMap, scanContext);
-                else
-                    PromoteCallerStack(frameAddress, fmf.TransitionBlockPtr, scanContext);
-                break;
-            }
-
-            case FrameType.DynamicHelperFrame:
-            {
-                Data.FramedMethodFrame fmf = target.ProcessedData.GetOrAdd<Data.FramedMethodFrame>(frameAddress);
-                Data.DynamicHelperFrame dhf = target.ProcessedData.GetOrAdd<Data.DynamicHelperFrame>(frameAddress);
-                ScanDynamicHelperFrame(fmf.TransitionBlockPtr, dhf.DynamicHelperFrameFlags, scanContext);
-                break;
-            }
-
-            case FrameType.CallCountingHelperFrame:
-            case FrameType.PrestubMethodFrame:
-            {
-                Data.FramedMethodFrame fmf = target.ProcessedData.GetOrAdd<Data.FramedMethodFrame>(frameAddress);
-                PromoteCallerStack(frameAddress, fmf.TransitionBlockPtr, scanContext);
-                break;
-            }
-
-            case FrameType.HijackFrame:
-                // TODO(stackref): Implement HijackFrame scanning (X86 only with FEATURE_HIJACK)
-                break;
-
-            case FrameType.ProtectValueClassFrame:
-                // TODO(stackref): Implement ProtectValueClassFrame scanning
-                break;
-
-            default:
-                // Base Frame::GcScanRoots_Impl is a no-op for most frame types.
-                break;
-        }
-    }
-
-    /// <summary>
-    /// Decodes a GCRefMap bitstream and reports GC references in the transition block.
-    /// Port of native TransitionFrame::PromoteCallerStackUsingGCRefMap (frames.cpp).
-    /// </summary>
-    private void PromoteCallerStackUsingGCRefMap(
-        TargetPointer transitionBlock,
-        TargetPointer gcRefMapBlob,
-        GcScanContext scanContext)
-    {
-        GCRefMapDecoder decoder = new(target, gcRefMapBlob);
-
-        if (target.PointerSize == 4)
-            decoder.ReadStackPop();
-
-        while (!decoder.AtEnd)
-        {
-            int pos = decoder.CurrentPos;
-            GCRefMapToken token = decoder.ReadToken();
-            uint offset = OffsetFromGCRefMapPos(pos);
-            TargetPointer slotAddress = new(transitionBlock.Value + offset);
-
-            switch (token)
-            {
-                case GCRefMapToken.Skip:
-                    break;
-                case GCRefMapToken.Ref:
-                    scanContext.GCReportCallback(slotAddress, GcScanFlags.None);
-                    break;
-                case GCRefMapToken.Interior:
-                    scanContext.GCReportCallback(slotAddress, GcScanFlags.GC_CALL_INTERIOR);
-                    break;
-                case GCRefMapToken.MethodParam:
-                case GCRefMapToken.TypeParam:
-                    break;
-                case GCRefMapToken.VASigCookie:
-                    // TODO(stackref): Implement VASIG_COOKIE handling
-                    break;
-            }
-        }
-    }
-
-    /// <summary>
-    /// Scans GC roots for a DynamicHelperFrame based on its flags.
-    /// Port of native DynamicHelperFrame::GcScanRoots_Impl (frames.cpp).
-    /// </summary>
-    private void ScanDynamicHelperFrame(
-        TargetPointer transitionBlock,
-        int dynamicHelperFrameFlags,
-        GcScanContext scanContext)
-    {
-        const int DynamicHelperFrameFlags_ObjectArg = 1;
-        const int DynamicHelperFrameFlags_ObjectArg2 = 2;
-
-        Target.TypeInfo tbType = target.GetTypeInfo(DataType.TransitionBlock);
-        uint argRegOffset = (uint)tbType.Fields[nameof(Data.TransitionBlock.ArgumentRegistersOffset)].Offset;
-
-        if ((dynamicHelperFrameFlags & DynamicHelperFrameFlags_ObjectArg) != 0)
-        {
-            TargetPointer argAddr = new(transitionBlock.Value + argRegOffset);
-            scanContext.GCReportCallback(argAddr, GcScanFlags.None);
-        }
-
-        if ((dynamicHelperFrameFlags & DynamicHelperFrameFlags_ObjectArg2) != 0)
-        {
-            TargetPointer argAddr = new(transitionBlock.Value + argRegOffset + (uint)target.PointerSize);
-            scanContext.GCReportCallback(argAddr, GcScanFlags.None);
-        }
-    }
-
-    /// <summary>
-    /// Resolves the GCRefMap for a Frame with m_pIndirection.
-    /// Port of native FindGCRefMap (frames.cpp).
-    /// </summary>
-    private TargetPointer FindGCRefMap(TargetPointer indirection)
-    {
-        if (indirection == TargetPointer.Null)
-            return TargetPointer.Null;
-
-        // Resolve the module from the indirection address.
-        // Matches native GetGCRefMap which calls FindModuleForGCRefMap(m_pIndirection)
-        IExecutionManager eman = target.Contracts.ExecutionManager;
-        TargetPointer zapModule = eman.FindReadyToRunModule(indirection);
-        if (zapModule == TargetPointer.Null)
-            return TargetPointer.Null;
-
-        // Get the ReadyToRunInfo from the module
-        Data.Module module = target.ProcessedData.GetOrAdd<Data.Module>(zapModule);
-        if (module.ReadyToRunInfo == TargetPointer.Null)
-            return TargetPointer.Null;
-
-        Data.ReadyToRunInfo r2rInfo = target.ProcessedData.GetOrAdd<Data.ReadyToRunInfo>(module.ReadyToRunInfo);
-        if (r2rInfo.ImportSections == TargetPointer.Null || r2rInfo.NumImportSections == 0)
-            return TargetPointer.Null;
-
-        // Compute RVA = indirection - imageBase
-        ulong imageBase = r2rInfo.LoadedImageBase.Value;
-        if (indirection.Value < imageBase)
-            return TargetPointer.Null;
-        ulong diff = indirection.Value - imageBase;
-        if (diff > uint.MaxValue)
-            return TargetPointer.Null;
-        uint rva = (uint)diff;
-
-        // READYTORUN_IMPORT_SECTION layout:
-        //   IMAGE_DATA_DIRECTORY Section (VirtualAddress:4, Size:4) = 8 bytes
-        //   ReadyToRunImportSectionFlags Flags (2 bytes)
-        //   ReadyToRunImportSectionType Type (1 byte)
-        //   BYTE EntrySize (1 byte)
-        //   DWORD Signatures (4 bytes)
-        //   DWORD AuxiliaryData (4 bytes)
-        // Total: 20 bytes
-        const int ImportSectionSize = 20;
-        const int SectionVAOffset = 0;
-        const int SectionSizeOffset = 4;
-        const int EntrySizeOffset = 11;
-        const int AuxiliaryDataOffset = 16;
-
-        TargetPointer sectionsBase = r2rInfo.ImportSections;
-        for (uint i = 0; i < r2rInfo.NumImportSections; i++)
-        {
-            TargetPointer sectionAddr = new(sectionsBase.Value + i * ImportSectionSize);
-            uint sectionVA = target.Read<uint>(sectionAddr + SectionVAOffset);
-            uint sectionSize = target.Read<uint>(sectionAddr + SectionSizeOffset);
-
-            if (rva >= sectionVA && rva < sectionVA + sectionSize)
-            {
-                byte entrySize = target.Read<byte>(sectionAddr + EntrySizeOffset);
-                if (entrySize == 0)
-                    return TargetPointer.Null;
-
-                uint index = (rva - sectionVA) / entrySize;
-                uint auxDataRva = target.Read<uint>(sectionAddr + AuxiliaryDataOffset);
-                if (auxDataRva == 0)
-                    return TargetPointer.Null;
-
-                TargetPointer gcRefMapBase = new(imageBase + auxDataRva);
-
-                // GCRefMap starts with a lookup index for stride-based access.
-                // GCREFMAP_LOOKUP_STRIDE is 1024 in the native code.
-                const uint GCREFMAP_LOOKUP_STRIDE = 1024;
-                uint lookupIndex = index / GCREFMAP_LOOKUP_STRIDE;
-                uint remaining = index % GCREFMAP_LOOKUP_STRIDE;
-
-                // Read the offset from the lookup table (array of DWORDs)
-                uint lookupOffset = target.Read<uint>(new TargetPointer(gcRefMapBase.Value + lookupIndex * 4));
-                TargetPointer p = new(gcRefMapBase.Value + lookupOffset);
-
-                // Linear scan past 'remaining' entries
-                while (remaining > 0)
-                {
-                    // Each entry is a variable-length sequence of bytes where the high bit
-                    // indicates continuation. Skip until we find a byte without the high bit set.
-                    while ((target.Read<byte>(p) & 0x80) != 0)
-                        p = new(p.Value + 1);
-                    p = new(p.Value + 1); // skip the final byte of this entry
-
-                    remaining--;
-                }
-
-                return p;
-            }
-        }
-
-        return TargetPointer.Null;
-    }
-
-    /// <summary>
-    /// Entry point for promoting caller stack GC references via method signature.
-    /// Matches native TransitionFrame::PromoteCallerStack (frames.cpp:1494).
-    /// </summary>
-    private void PromoteCallerStack(
-        TargetPointer frameAddress,
-        TargetPointer transitionBlock,
-        GcScanContext scanContext)
-    {
-        Data.FramedMethodFrame fmf = target.ProcessedData.GetOrAdd<Data.FramedMethodFrame>(frameAddress);
-        TargetPointer methodDescPtr = fmf.MethodDescPtr;
-        if (methodDescPtr == TargetPointer.Null)
-            return;
-
-        ReadOnlySpan<byte> signature;
-        try
-        {
-            signature = GetMethodSignatureBytes(methodDescPtr);
-        }
-        catch (System.Exception)
-        {
-            return;
-        }
-
-        if (signature.IsEmpty)
-            return;
-
-        MethodSignature<GcTypeKind> methodSig;
-        try
-        {
-            unsafe
-            {
-                fixed (byte* pSig = signature)
-                {
-                    BlobReader blobReader = new(pSig, signature.Length);
-                    SignatureDecoder<GcTypeKind, object?> decoder = new(
-                        GcSignatureTypeProvider.Instance, metadataReader: null!, genericContext: null);
-                    methodSig = decoder.DecodeMethodSignature(ref blobReader);
-                }
-            }
-        }
-        catch (System.Exception)
-        {
-            // If signature decoding fails (e.g., ELEMENT_TYPE_INTERNAL), skip this frame.
-            // The GCRefMap path handles these cases when available.
-            return;
-        }
-
-        if (methodSig.Header.CallingConvention is SignatureCallingConvention.VarArgs)
-        {
-            // TODO(stackref): VarArg path — read VASigCookie from frame
-            return;
-        }
-
-        IRuntimeTypeSystem rts = target.Contracts.RuntimeTypeSystem;
-        MethodDescHandle mdh = rts.GetMethodDescHandle(methodDescPtr);
-
-        bool hasThis = methodSig.Header.IsInstance;
-        bool hasRetBuf = methodSig.ReturnType is GcTypeKind.Other;
-        bool requiresInstArg = false;
-        bool isAsync = false;
-        bool isValueTypeThis = false;
-
-        try
-        {
-            requiresInstArg = rts.RequiresInstArg(mdh);
-            isAsync = rts.IsAsyncMethod(mdh);
-
-            // TODO(stackref): Detect value type 'this' (needs IRuntimeTypeSystem.IsValueType)
-            // TODO(stackref): String constructor clears HasThis
-        }
-        catch
-        {
-        }
-
-        PromoteCallerStackHelper(transitionBlock, methodSig, hasThis, hasRetBuf,
-            requiresInstArg, isAsync, isValueTypeThis, scanContext);
-    }
-
-    /// <summary>
-    /// Core logic for promoting caller stack GC references.
-    /// Matches native TransitionFrame::PromoteCallerStackHelper (frames.cpp:1560).
-    /// </summary>
-    private void PromoteCallerStackHelper(
-        TargetPointer transitionBlock,
-        MethodSignature<GcTypeKind> methodSig,
-        bool hasThis,
-        bool hasRetBuf,
-        bool requiresInstArg,
-        bool isAsync,
-        bool isValueTypeThis,
-        GcScanContext scanContext)
-    {
-        int numRegistersUsed = 0;
-        if (hasThis)
-            numRegistersUsed++;
-        if (hasRetBuf)
-            numRegistersUsed++;
-        if (requiresInstArg)
-            numRegistersUsed++;
-        if (isAsync)
-            numRegistersUsed++;
-
-        bool isArm64 = IsTargetArm64();
-        if (isArm64)
-            numRegistersUsed++;
-
-        if (hasThis)
-        {
-            int thisPos = isArm64 ? 1 : 0;
-            uint thisOffset = OffsetFromGCRefMapPos(thisPos);
-            TargetPointer thisAddr = new(transitionBlock.Value + thisOffset);
-            GcScanFlags thisFlags = isValueTypeThis ? GcScanFlags.GC_CALL_INTERIOR : GcScanFlags.None;
-            scanContext.GCReportCallback(thisAddr, thisFlags);
-        }
-
-        // TODO(stackref): Promote async continuation pointer at its specific offset
-
-        int pos = numRegistersUsed;
-        foreach (GcTypeKind kind in methodSig.ParameterTypes)
-        {
-            uint offset = OffsetFromGCRefMapPos(pos);
-            TargetPointer slotAddress = new(transitionBlock.Value + offset);
-
-            switch (kind)
-            {
-                case GcTypeKind.Ref:
-                    scanContext.GCReportCallback(slotAddress, GcScanFlags.None);
-                    break;
-                case GcTypeKind.Interior:
-                    scanContext.GCReportCallback(slotAddress, GcScanFlags.GC_CALL_INTERIOR);
-                    break;
-                case GcTypeKind.Other:
-                    // TODO(stackref): Value type GCDesc scanning
-                    break;
-                case GcTypeKind.None:
-                    break;
-            }
-            pos++;
-        }
-    }
-
-    private ReadOnlySpan<byte> GetMethodSignatureBytes(TargetPointer methodDescPtr)
-    {
-        IRuntimeTypeSystem rts = target.Contracts.RuntimeTypeSystem;
-        MethodDescHandle mdh = rts.GetMethodDescHandle(methodDescPtr);
-
-        if (rts.IsStoredSigMethodDesc(mdh, out ReadOnlySpan<byte> storedSig))
-            return storedSig;
-
-        uint methodToken = rts.GetMethodToken(mdh);
-        if (methodToken == 0x06000000)
-            return default;
-
-        TargetPointer methodTablePtr = rts.GetMethodTable(mdh);
-        TypeHandle typeHandle = rts.GetTypeHandle(methodTablePtr);
-        TargetPointer modulePtr = rts.GetModule(typeHandle);
-
-        ILoader loader = target.Contracts.Loader;
-        ModuleHandle moduleHandle = loader.GetModuleHandleFromModulePtr(modulePtr);
-
-        IEcmaMetadata ecmaMetadata = target.Contracts.EcmaMetadata;
-        MetadataReader? mdReader = ecmaMetadata.GetMetadata(moduleHandle);
-        if (mdReader is null)
-            return default;
-
-        MethodDefinitionHandle methodDefHandle = MetadataTokens.MethodDefinitionHandle((int)(methodToken & 0x00FFFFFF));
-        MethodDefinition methodDef = mdReader.GetMethodDefinition(methodDefHandle);
-        BlobReader blobReader = mdReader.GetBlobReader(methodDef.Signature);
-        return blobReader.ReadBytes(blobReader.Length);
-    }
-
-    private uint OffsetFromGCRefMapPos(int pos)
-    {
-        Target.TypeInfo tbType = target.GetTypeInfo(DataType.TransitionBlock);
-        uint firstSlotOffset = (uint)tbType.Fields[nameof(Data.TransitionBlock.FirstGCRefMapSlot)].Offset;
-        return firstSlotOffset + (uint)(pos * target.PointerSize);
-    }
-
-    private bool IsTargetArm64()
-    {
-        return target.Contracts.RuntimeInfo.GetTargetArchitecture() is RuntimeInfoArchitecture.Arm64;
     }
 }
